Backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
pydantic==1.10.2
python-dotenv==1.0.0
twilio==9.0.1
httpx==0.24.0
pytest==7.4.0
pytest-asyncio==0.21.0
